tests/test_phase2_files.py
"""Tests for Phase 2 - Files module (upload, parsing, chunking, embedding)."""

import pytest
import io
from uuid import uuid4
from typing import AsyncGenerator
from unittest.mock import patch, MagicMock

from fastapi.testclient import TestClient
from sqlalchemy.ext.asyncio import AsyncSession

from ragcore.main import app
from ragcore.db.database import get_db_session
from ragcore.models import File, Chunk, Job
from ragcore.modules.files.parser import FileParser
from ragcore.modules.files.chunker import TextChunker
from ragcore.modules.files.pipeline import FileProcessingPipeline

# Test client
client = TestClient(app)


# ============================================================================
# Parser Tests
# ============================================================================


class TestFileParser:
    """Tests for PDF and DOCX parsing."""

    def test_parse_pdf_simple(self):
        """Test parsing a simple PDF."""
        # Create minimal PDF bytes (using PyMuPDF)
        try:
            import fitz
            doc = fitz.open()
            page = doc.new_page()
            page.insert_text((50, 50), "Hello World")

            pdf_bytes = io.BytesIO()
            doc.save(pdf_bytes)
            pdf_bytes.seek(0)

            text, metadata = FileParser.parse_pdf(pdf_bytes.getvalue())

            assert "Hello World" in text
            assert metadata["file_type"] == "pdf"
            assert metadata["page_count"] == 1
            assert metadata["char_count"] > 0

        except ImportError:
            pytest.skip("PyMuPDF not installed")

    def test_parse_docx_simple(self):
        """Test parsing a simple DOCX."""
        try:
            from docx import Document
            from docx.shared import Pt

            doc = Document()
            doc.add_paragraph("Hello World")

            docx_bytes = io.BytesIO()
            doc.save(docx_bytes)
            docx_bytes.seek(0)

            text, metadata = FileParser.parse_docx(docx_bytes.getvalue())

            assert "Hello World" in text
            assert metadata["file_type"] == "docx"
            assert metadata["paragraph_count"] >= 1
            assert metadata["char_count"] > 0

        except ImportError:
            pytest.skip("python-docx not installed")

    def test_parse_invalid_pdf(self):
        """Test error handling for corrupted PDF."""
        invalid_bytes = b"This is not a PDF"

        with pytest.raises(ValueError, match="PDF parsing error"):
            FileParser.parse_pdf(invalid_bytes)

    def test_parse_router_pdf(self):
        """Test parse router with PDF content type."""
        try:
            import fitz
            doc = fitz.open()
            page = doc.new_page()
            page.insert_text((50, 50), "Test content")

            pdf_bytes = io.BytesIO()
            doc.save(pdf_bytes)
            pdf_bytes.seek(0)

            text, metadata = FileParser.parse(
                pdf_bytes.getvalue(),
                "application/pdf"
            )

            assert "Test content" in text
            assert metadata["file_type"] == "pdf"

        except ImportError:
            pytest.skip("PyMuPDF not installed")


# ============================================================================
# Chunker Tests
# ============================================================================


class TestTextChunker:
    """Tests for text chunking with token awareness."""

    def test_chunker_initialization(self):
        """Test chunker initialization."""
        chunker = TextChunker(
            chunk_size_tokens=512,
            chunk_overlap_tokens=50
        )

        assert chunker.chunk_size_tokens == 512
        assert chunker.chunk_overlap_tokens == 50

    def test_chunker_simple_text(self):
        """Test chunking simple text."""
        chunker = TextChunker(chunk_size_tokens=100, chunk_overlap_tokens=10)

        text = "This is a test. " * 20  # Create larger text
        chunks = chunker.chunk(text)

        assert len(chunks) > 0
        for chunk in chunks:
            assert "text" in chunk
            assert "tokens" in chunk
            assert "chunk_index" in chunk
            assert chunk["tokens"] > 0

    def test_chunker_with_metadata(self):
        """Test chunking with metadata."""
        chunker = TextChunker()

        text = "Sample document text. " * 10
        metadata = {"source": "test.pdf", "author": "Test"}

        chunks = chunker.chunk(text, metadata=metadata)

        for chunk in chunks:
            assert "metadata" in chunk
            assert chunk["metadata"]["source"] == "test.pdf"

    def test_chunker_overlap(self):
        """Test that chunks have overlap."""
        chunker = TextChunker(
            chunk_size_tokens=50,
            chunk_overlap_tokens=10
        )

        text = "word " * 100  # 100 words
        chunks = chunker.chunk(text)

        assert len(chunks) > 1
        # Later chunks should have overlap with earlier ones
        if len(chunks) >= 2:
            first_chunk_end = chunks[0]["text"].split()[-5:]
            second_chunk_start = chunks[1]["text"].split()[:5]
            # There should be some overlap in the word sequences


# ============================================================================
# File Upload Endpoint Tests
# ============================================================================


class TestFileUploadEndpoint:
    """Tests for file upload HTTP endpoint."""

    def test_upload_file_invalid_type(self):
        """Test upload with invalid file type."""
        response = client.post(
            "/files/upload",
            files={"file": ("test.txt", b"plain text", "text/plain")}
        )

        # TBD: implementation may vary
        # Could be 400 or 422 depending on validation
        assert response.status_code in [400, 422]

    def test_upload_file_valid_pdf_mock(self):
        """Test upload with mock PDF (avoids dependencies)."""
        pdf_bytes = b"%PDF-1.4\n..."  # Minimal PDF header

        with patch('ragcore.modules.files.router.queue_file_processing_job'):
            # Note: This test would fail without proper PDF
            # In production, use real PDF or mock the parser
            pass

    def test_upload_file_size_limit(self):
        """Test upload exceeds size limit."""
        # Create file larger than max_file_size_mb
        large_data = b"x" * (100 * 1024 * 1024)  # 100MB

        response = client.post(
            "/files/upload",
            files={"file": ("large.pdf", large_data, "application/pdf")}
        )

        assert response.status_code == 400


# ============================================================================
# File Status Tests
# ============================================================================


class TestFileStatusTransitions:
    """Tests for file status state machine."""

    @pytest.mark.asyncio
    async def test_file_status_pending(self):
        """Test file created with pending status."""
        async with get_db_session() as session:
            file = File(
                filename="test.pdf",
                file_size=1024,
                content_type="application/pdf",
                status="pending"
            )
            session.add(file)
            await session.commit()

            assert file.status == "pending"
            assert file.chunks_count == 0

    @pytest.mark.asyncio
    async def test_file_cascade_delete(self):
        """Test file deletion cascades to chunks."""
        from sqlalchemy import select

        async with get_db_session() as session:
            # Create file with chunks
            file = File(
                filename="test.pdf",
                file_size=1024,
                content_type="application/pdf",
                status="ready",
                chunks_count=2
            )
            session.add(file)
            await session.flush()

            chunk1 = Chunk(
                file_id=file.id,
                chunk_index=0,
                text="Sample text",
                tokens=10,
                embedding=None,
                metadata_={}
            )
            chunk2 = Chunk(
                file_id=file.id,
                chunk_index=1,
                text="More text",
                tokens=8,
                embedding=None,
                metadata_={}
            )
            session.add_all([chunk1, chunk2])
            await session.commit()

            file_id = file.id

            # Delete file
            await session.delete(file)
            await session.commit()

            # Verify chunks are deleted
            stmt = select(Chunk).where(Chunk.file_id == file_id)
            result = await session.execute(stmt)
            remaining = result.scalars().all()

            assert len(remaining) == 0


# ============================================================================
# Pipeline Tests
# ============================================================================


class TestFileProcessingPipeline:
    """Tests for the file processing pipeline."""

    @pytest.mark.asyncio
    async def test_pipeline_initialization(self):
        """Test pipeline initializes correctly."""
        pipeline = FileProcessingPipeline()

        assert pipeline.chunker is not None
        assert pipeline.embedding_provider is not None
        assert pipeline.embedding_model is not None

    @pytest.mark.asyncio
    async def test_pipeline_parse_stage(self):
        """Test pipeline parse stage."""
        with patch('ragcore.modules.files.parser.FileParser.parse') as mock_parse:
            mock_parse.return_value = ("Sample text", {"file_type": "pdf"})

            # Create test file
            async with get_db_session() as session:
                file = File(
                    filename="test.pdf",
                    file_size=100,
                    content_type="application/pdf",
                    status="pending"
                )
                session.add(file)
                await session.commit()
                file_id = file.id

            # Pipeline would normally update status
            # Verify file can be fetched
            async with get_db_session() as session:
                fetched = await session.get(File, file_id)
                assert fetched is not None


# ============================================================================
# Chunk Retrieval Tests
# ============================================================================


class TestChunkRetrieval:
    """Tests for retrieving file chunks."""

    def test_get_file_chunks_not_found(self):
        """Test getting chunks for non-existent file."""
        fake_id = str(uuid4())
        response = client.get(f"/files/{fake_id}/chunks")

        assert response.status_code == 404

    def test_delete_file(self):
        """Test file deletion endpoint."""
        fake_id = str(uuid4())
        response = client.delete(f"/files/{fake_id}")

        assert response.status_code == 404


# ============================================================================
# Files List Tests
# ============================================================================


class TestFilesListEndpoint:
    """Tests for listing files."""

    def test_list_files_empty(self):
        """Test listing files when none exist."""
        response = client.get("/files")

        assert response.status_code == 200
        assert isinstance(response.json(), list)


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
